Refactor further card_patcher.c and card_finder.c
• Split card_patcher.c into card_patcher_common.c, card_patcher_arm7.c and card_patcher_arm9.c
• Split card_finder.c into card_finder_common.c, card_finder_arm7.c and card_finder_arm9.c
• Make card_patcher_*.c and card_finder_*.c not have extern values anymore
• Split finders into “normal” and “thumb”
• Split normal card read finders into read type “0” and “1”
• Make finders return pointers
• Make function “generateA7InstrThumb” return the instructions, similarly to “generateA7Instr”
• Fix all compiler warnings
