+++ bootloader/source/hook.c
@@ -61,6 +61,7 @@ static const u32 handlerEndSig[4] = {
 	0xe12fff10		// bx   r0
 };
 
+/*
 // interruptDispatcher.s jump_intr:
 static const u32 homebrewSig[5] = {
 	0xE5921000, // ldr    r1, [r2]        @ user IRQ handler address
@@ -117,10 +118,11 @@ static const u32 homebrewAccelSig2010Patched[4] = {
 	0x22088819   , // ...
 	0x0412430A   , // ...
 };
+*/
 
 static const int MAX_HANDLER_SIZE = 50;
 
-static u32* hookInterruptHandlerHomebrew (u32* addr, size_t size) {
+/*static u32* hookInterruptHandlerHomebrew (u32* addr, size_t size) {
 	u32* end = addr + size/sizeof(u32);
 
 	// Find the start of the handler
@@ -208,7 +210,7 @@ static u32* hookAccelIPCHomebrew2010 (u32* addr, size_t size) {
 
 	// The first entry in the table is for the Vblank handler, which is what we want
 	return addr;
-}
+}*/
 
 static u32* hookInterruptHandler (u32* addr, size_t size) {
 	u32* end = addr + size/sizeof(u32);
