+++ bootloader/source/save_patch_v1.c
@@ -4,7 +4,7 @@ u32 savePatchV1(const tNDSHeader* ndsHeader, u32* cardEngineLocation, module_par
 	dbg_printf("\nArm7 (patch v1.0)\n");
 
 	// Find the relocation signature
-	u32 relocationStart = getOffset(
+	u32 relocationStart = (u32)findOffset(
 		(u32*)ndsHeader->arm7destination, ndsHeader->arm7binarySize,
 		relocateStartSignature, 1,
 		1
@@ -28,7 +28,7 @@ u32 savePatchV1(const tNDSHeader* ndsHeader, u32* cardEngineLocation, module_par
 		dbg_printf("Error in relocation checking method 1\n");
 		
 		// Found the beginning of the next function
-		u32 nextFunction = getOffset(
+		u32 nextFunction = (u32)findOffset(
 			(u32*)relocationStart, ndsHeader->arm7binarySize,
 			nextFunctiontSignature, 1,
 			1
@@ -70,7 +70,7 @@ u32 savePatchV1(const tNDSHeader* ndsHeader, u32* cardEngineLocation, module_par
 	dbg_printf("\n");
 
 	// Find the card read
-	u32 cardReadEndAddr = getOffset(
+	u32 cardReadEndAddr = (u32)findOffset(
 		(u32*)ndsHeader->arm7destination, 0x00400000, 
 		a7cardReadSignature, 2,
 		1
@@ -89,7 +89,7 @@ u32 savePatchV1(const tNDSHeader* ndsHeader, u32* cardEngineLocation, module_par
 	dbg_hexa(cardstructAddr);
 	dbg_printf("\n");
 
-	u32 readCacheEnd = getOffset(
+	u32 readCacheEnd = (u32)findOffset(
 		(u32*)cardReadEndAddr, 0x18000 - cardReadEndAddr,
 		&cardstructAddr, 1,
 		1
@@ -116,7 +116,7 @@ u32 savePatchV1(const tNDSHeader* ndsHeader, u32* cardEngineLocation, module_par
 	// if out of specific ram range...
 	if (specificWramAddr < 0x37F8000 || specificWramAddr > 0x380FFFF) {
 		dbg_printf("Retry the search\n");
-		JumpTableFunc = getOffset(
+		JumpTableFunc = (u32)findOffset(
 			(u32*)JumpTableFunc, 0x18000 - JumpTableFunc,
 			&cardstructAddr, 1,
 			1
@@ -134,7 +134,7 @@ u32 savePatchV1(const tNDSHeader* ndsHeader, u32* cardEngineLocation, module_par
 	dbg_hexa(specificWramAddr);
 	dbg_printf("\n");
 
-	u32 someAddr_799C = getOffset(
+	u32 someAddr_799C = (u32)findOffset(
 		(u32*)ndsHeader->arm7destination, 0x18000,
 		a7something2Signature, 2,
 		1
